improve(engine): improve telemetry packaging
