+++ engine/src/main/java/org/camunda/bpm/engine/impl/telemetry/dto/Internals.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.impl.telemetry.node;
+package org.camunda.bpm.engine.impl.telemetry.dto;
 
 public class Internals {
